merge addProduct refactor to production

+++ src/main/java/com/chompfooddeliveryapp/payload/ProductSummary.java
@@ -16,5 +16,5 @@ public class ProductSummary {
     private String productOwner;
     private String productImage;
     private Integer productQuantity;
-    private Long productPrice;
+    private Double productPrice;
 }
